Constructive options for class 'POSIXct'

These options will be used on objects of class 'POSIXct'.

Usage

opts_POSIXct(
  constructor = c("as.POSIXct", ".POSIXct", "as_datetime", "as.POSIXct.numeric",
    "as_datetime.numeric", "next", "atomic"),
  ...,
  origin = "1970-01-01"
)

Arguments

constructor

String. Name of the function used to construct the environment, see Details section.

...

Should not be used. Forces passing arguments by name.

origin

Origin to be used, ignored when irrelevant.

Value

An object of class <constructive_options/constructive_options_factor>

Details

Depending on constructor, we construct the environment as follows:

"as.POSIXct" (default): Build the object using a as.POSIXct() call on a character vector.

".POSIXct" : Build the object using a .POSIXct() call on a numeric vector.

"as_datetime" : Build the object using a lubridate::as_datetime() call on a character vector.

"next" : Use the constructor for the next supported class. Call .class2() on the object to see in which order the methods will be tried.

"atomic" : We define as an atomic vector and repair attributes.