Constructive options for type 'environment'

Constructive options for type 'environment'

Usage

opts_environment(
  constructor = c("list2env", "as.environment", "new.env", "topenv", "new_environment"),
  ...,
  recurse = FALSE
)

Arguments

constructor

String. Name of the function used to construct the environment, see Details section.

...

Should not be used. Forces passing arguments by name.

recurse

Boolean. Only considered if constructor is "list2env" or "new_environment". Whether to attempt to recreate all parent environments until a known environment is found, if FALSE (the default) we will use topenv() to find a known ancestor to set as the parent.

Value

An object of class <constructive_options/constructive_options_environment>

Details

Environments use reference semantics, they cannot be copied. An attempt to copy an environment would indeed yield a different environment and identical(env, copy) would be FALSE.
Moreover every environment (except for emptyenv()) has a parent and thus to copy the environment we'd have to have a way to point to the parent, or copy it too.
For this reason environments are constructive's cryptonite. They make some objects impossible to reproduce exactly. And since every function or formula has one they're hard to avoid
Luckily in many cases we don't need a perfect copy of an environment to reproduce an object faithfully from a practical standpoint, and in a some cases we might build code that points to a specific environment such as .GlobalEnv or a given namespace.
constructive will not signal any difference if it can reproduce an equivalent environment, defined as containing the same values and having a same or equivalent parent.
See also the ignore_function_env argument in ?construct, which disables the check of environments of function.

Environment construction works as follows:

Special environments .GlobalEnv, .BaseNamespaceEnv, baseenv() and emptyenv() are used whenever possible.

The idiom asNamespace("pkg") is used whenever possible.

The idiom as.environment("package:pkg") is used whenever possible.

The constructor argument is used to recreate a similar environment whenever the above conditions are not met.

Depending on constructor, we construct the environment as follows:

"list2env" (default): We construct the environment as a list then use base::list2env() to convert it to an environment and assign it a parent. By default we will use base::topenv() to construct a parent. If recurse is TRUE the parent will be built recursively so all ancestors will be created until we meet a known environment, this might be verbose and will fail if environments are nested too deep or have a circular relationship, but if it works it's the most faithful method. If the environment is empty we use new.env(parent=) for a more economic syntax.

"new_environment" : Similar to the above, but using rlang::new_environment().

"new.env" : All environments will be recreated with the code "base::new.env()", effectively creating an empty environment child of the local (often global) environment. This is enough in cases where the environment doesn't matter at all (or matters as long as it inherits from the local environment), this is the case for most formulas. recurse is ignored.

"as.environment" : we attempt to construct the environment as a list and use base::as.environment() on top of it, as in as.environment(list(a=1, b=2)), it will contain the same variables as the original environment but the parent will be the emptyenv(). recurse is ignored.

"topenv" : we construct base::topenv(x), see ?topenv. recurse is ignored.