Constructive options for matrices

Matrices are atomic vectors, lists, or objects of type "expression" with a "dim" attributes of length 2.

Usage

opts_matrix(constructor = c("matrix", "array", "cbind", "rbind", "next"), ...)

Arguments

constructor

String. Name of the function used to construct the object.

...

Additional options used by user defined constructors through the opts object

Value

An object of class <constructive_options/constructive_options_matrix>

Details

Depending on constructor, we construct the object as follows:

"matrix" : We use matrix()

"array" : We use array()

"cbind","rbind" : We use cbind() or "rbind()", this makes named columns and rows easier to read.

"next" : Use the constructor for the next supported class. Call .class2() on the object to see in which order the methods will be tried. This will usually be equivalent to "array"

"atomic" : We define as an atomic vector and repair attributes